refactor: rename folder to atoms, share h1 style
